requirements.txt
asgiref==3.5.2
awsebcli==3.20.3
botocore==1.23.54
cement==2.8.2
certifi==2022.6.15
charset-normalizer==2.0.12
colorama==0.4.3
distlib==0.3.5
Django==3.2.15
django-appconf==1.0.5
django-ckeditor==6.3.2
django-crispy-forms==1.14.0
django-environ==0.9.0
django-js-asset==2.0.0
django-mobile==0.7.0
django-phonenumber-field==6.3.0
django-responsive==0.3.0
django-rest-framework==0.1.0
django-userforeignkey==0.5.0
djangorestframework==3.13.1
environ==1.0
filelock==3.8.0
future==0.16.0
idna==3.3
importlib-metadata==4.12.0
jmespath==0.10.0
pathspec==0.9.0
phonenumbers==8.12.52
Pillow==9.1.1
platformdirs==2.5.2
psycopg2-binary==2.9.3
pyodbc==4.0.34
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==5.4.1
requests==2.26.0
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.2
termcolor==1.1.0
typing_extensions==4.3.0
urllib3==1.26.12
virtualenv==20.16.3
wcwidth==0.1.9
zipp==3.8.1
